d05/requirements.txt
django==5.1.1
psycopg2-binary==2.9.9
python-decouple==3.8
